completed user check for each route

+++ routes/habits.js
@@ -11,12 +11,12 @@ MongoClient.connect("mongodb://localhost/HabitTracker", { useUnifiedTopology: tr
 
     //Get all users
     router.get('/all', (req, res) => {
-        usersCollection.find().toArray() //Read a document in the database
-        .then(results => {
-          (result) ? res.send(result) //If one or more documents exist 
-          : res.send("No documents"); //If there are no documents
-          })
-        .catch(error => console.error(error))
+      usersCollection.find().toArray() //Read a document in the database
+      .then(results => {
+        (result) ? res.send(result) //If one or more documents exist 
+        : res.send("No documents"); //If there are no documents
+        })
+      .catch(error => console.error(error))
     });
     
     //Get a single user
@@ -32,38 +32,38 @@ MongoClient.connect("mongodb://localhost/HabitTracker", { useUnifiedTopology: tr
     
     //Add a new user
     router.post('/add-user', (req, res) => {
-        req.body.username = req.body.username.toLowerCase(); 
-        usersCollection.findOne({"username": {$eq:req.body.username}}) //Read a document in the database
-        .then(result => {
-          if (result) { //If the document does exist
-            res.send("User already exists");
-          } else { //If the document does not exist  
-            usersCollection.insertOne(req.body) //Create a document in the database
-            .then(result => {
-              res.send("User added");
-            })
-            .catch(error => res.status(500).send(error)); 
-          };
-        })
-        .catch(error => res.status(500).send(error)); 
+      req.body.username = req.body.username.toLowerCase(); 
+      usersCollection.findOne({"username": {$eq:req.body.username}}) //Read a document in the database
+      .then(result => {
+        if (result) { //If the document does exist
+          res.send("User already exists");
+        } else { //If the document does not exist  
+          usersCollection.insertOne(req.body) //Create a document in the database
+          .then(result => {
+            res.send("User added");
+          })
+          .catch(error => res.status(500).send(error)); 
+        };
+      })
+      .catch(error => res.status(500).send(error)); 
     });
 
-    //Add a new habit to a user**
+    //Add a new habit to a user
     router.put('/add-habit/:username', (req, res) => {
-        req.params.username = req.params.username.toLowerCase();
-        usersCollection.findOne({"username": {$eq:req.body.username}}) //Read a document in the database
-        .then(result => {
-          if (result) { //If the document does exist
-            usersCollection.updateOne({"username": req.params.username}, {$push: {"habit": req.body}}, true, false) //Update a document in the database
-            .then(
-                res.send()
-            )
-            .catch(error => res.status(500).send(error)); 
-          } else { //If the document does not exist
-            res.send("User does not exist"); 
-          };
-        })
-        .catch(error => res.status(500).send(error)); 
+      req.params.username = req.params.username.toLowerCase();
+      usersCollection.findOne({"username": {$eq:req.params.username}}) //Read a document in the database
+      .then(result => {
+        if (result) { //If the document does exist
+          usersCollection.updateOne({"username": req.params.username}, {$push: {"habit": req.body}}, true, false) //Update a document in the database
+          .then(
+            res.send("Habit updated")
+          )
+          .catch(error => res.status(500).send(error)); 
+        } else { //If the document does not exist
+          res.send("User does not exist"); 
+        };
+      })
+      .catch(error => res.status(500).send(error)); 
     });
 
     //Update a new habit to a user - unfinished
@@ -77,42 +77,62 @@ MongoClient.connect("mongodb://localhost/HabitTracker", { useUnifiedTopology: tr
     .catch(error => console.error(error))
     }); */
 
-    // tracking route - Done?
+    //Update a tracking field
     router.put('/update-habit/:username/:habitID/:indexTracking/:trueOrFalse', (req, res) => {
-        req.params.username = req.params.username.toLowerCase();
-      usersCollection.updateOne({ "username": req.params.username },
-      {$set: { [`habit.${req.params.habitID}.tracking.${req.params.indexTracking}`] : eval(req.params.trueOrFalse)}}, true,false)
-      .then(
-          res.send("Tracking updated")
-      )
-      .catch(error => console.error(error))
-      });
+      req.params.username = req.params.username.toLowerCase();
+      usersCollection.findOne({"username": {$eq:req.params.username}}) //Read a document in the database
+      .then(result => {
+        if (result) { //If the document does exist
+          usersCollection.updateOne({ "username": req.params.username },
+          {$set: { [`habit.${req.params.habitID}.tracking.${req.params.indexTracking}`] : eval(req.params.trueOrFalse)}}, true,false) //Update a document in the database
+          .then(
+            res.send("Tracking field updated")
+          )
+          .catch(error => res.status(500).send(error));
+        } else { //If the document does not exist
+          res.send("User does not exist"); 
+        };
+      })
+      .catch(error => res.status(500).send(error));  
+    });
 
-    //Delete a habit of a user*
+    //Delete a habit of a user
     router.put('/delete-habit/:username/:habitID', (req, res) => {
         req.params.username = req.params.username.toLowerCase();
-        console.log(req.params.habitID)
-        usersCollection.updateOne({"username": req.params.username}, {$unset: {[`habit.${req.params.habitID}`] : 1}})
-        .then(
-            console.log("unset")
-        )
-        usersCollection.updateOne({"username": req.params.username}, {$pull : {"habit" : null}})
-        .then(
-            res.send("Deleted a habit from user")
-        )
-        .catch(error => console.error(error))
-    });
+        usersCollection.findOne({"username": {$eq:req.params.username}}) //Read a document in the database
+        .then(result => {
+          if (result) { //If the document does exist
+            usersCollection.updateOne({"username": req.params.username}, {$unset: {[`habit.${req.params.habitID}`] : 1}}) //Update a document in the database
+            .then(
+            usersCollection.updateOne({"username": req.params.username}, {$pull : {"habit" : null}})) //Update a document in the database
+            .then(
+                res.send("Deleted a habit from user")
+            )
+            .catch(error => res.status(500).send(error));
+          } else { //If the document does not exist
+            res.send("User does not exist"); 
+          };
+        })
+        .catch(error => res.status(500).send(error));  
+      });
     
-    //Delete a user*
+    //Delete a user
       router.delete('/delete-user/:username', (req, res) => {
         req.params.username = req.params.username.toLowerCase();
-        usersCollection.deleteOne({"username": {$eq: req.params.username}})
-        .then(
-            res.send("Deleted a user from the database")
-        )
-        .catch(error => console.error(error))
-    });
-  });
-
+        usersCollection.findOne({"username": {$eq:req.params.username}}) //Read a document in the database
+        .then(result => {
+        if (result){
+          usersCollection.deleteOne({"username": {$eq: req.params.username}})
+          .then(
+            res.send("Deleted user")
+          )
+        }
+        else{
+          res.send("User does not exist")
+        }
+      })
+      .catch(error => res.status(500).send(error))
+    })
+  })
 
 module.exports = router;
